+++ core/src/main/kotlin/sift/core/tree/TreeDsl.kt
@@ -1,5 +1,6 @@
 package sift.core.tree
 
+import net.onedaybeard.collectionsby.anyBy
 import sift.core.entity.Entity
 
 class TreeDsl(private val hosted: Tree<EntityNode>) {